rename to huawei vrp
