+++ mkdocs.yml
@@ -43,6 +43,7 @@ nav:
               - FTOS10v: manual/kinds/vr-ftosv.md
               - Enterprise SONiC: manual/kinds/dell_sonic.md
           - MikroTik RouterOS: manual/kinds/vr-ros.md
+          - Huawei VRP: manual/kinds/huawei_vrp.md
           - IPInfusion OcNOS: manual/kinds/ipinfusion-ocnos.md
           - OpenBSD: manual/kinds/openbsd.md
           - FreeBSD: manual/kinds/freebsd.md
